lowering: split `finally` blocks for exceptional control-flow
This change duplicates `finally` blocks in lowered IR, so that they can
have a static nesting depth in the `try-catch` hierarchy.

Previously, a `finally` block had many non-excepting control-flow paths
and an exceptional control-flow path which passed through the same
block:

```
error   non-error
    \   /
     \ /
      |
    finally block
      |
     / \
    /   \
 error   non-error
```

The branching logic dynamically guaranteed that the `try/catch` blocks
are respected here (you never flow in an error path and out a non-error
one), but the compiler doesn't know that.

In a couple places, the compiler assumes that it can actually "color"
the CFG such that there is a static nesting depth at each BasicBlock
(i.e. each BasicBlock can be labeled w/ a unique enclosing `try` /
`catch` scope). The above `finally` pattern violates that assumption.

This was mostly benign as far as I can tell, but I want to extend the
lifetimes of our Event Handlers (`jl_handler_t`) until the end of a
`catch` block (rather than the start) which noticeably breaks
`llvm-lower-handlers.cpp`.

(@keno was very clear about this assumption in the comments for that pass.)