codegen: Restore `ct->scope` in `jl_eh_restore_state`
This eliminates the need to associate a `catch` with every
`with(...) do ... end` block, which was really just acting as a landing
pad to restore `jl_current_task->scope` in the majority of cases.

This change does not actually update lowering to remove the unnecessary
`catch` block - that's left as a follow-up.