Explicitly store a module's location
Revise wants to know what file a module's `module` definition is in.
Currently it does this by looking at the source location for the
implicitly generated `eval` method. This is terrible for two reasons:

1. The method may not exist if the module is a baremodule (which is not
   particularly common, which is probably why we haven't seen it).
2. The fact that the implicitly generated `eval` method has this location
   information is an implementation detail that I'd like to get rid of
   (#55949).

This PR adds explicit file/line info to `Module`, so that Revise doesn't
have to use the hack anymore.