Remove unreachable error branch in memset calls
Some places use the pattern memset(A, v, length(A)), which requires a conversion
UInt(length(A)). This is technically fallible, but can't actually fail when A
is a Memory or Array.
Remove the dead error branch by casting to UInt instead.

Similarly, in repeat(x, r), r is first checked to be nonnegative, then converted
to UInt, then used in multiple calls where it is converted to UInt each time.
Here, only do it once.