Document type-piracy / type-leakage restrictions for `require_stdlib`
As a recent offender, this tries to lay down some guidelines for the
behavior that stdlibs and the callers of `require_stdlib` must adhere
too to avoid "duplicate stdlib" bugs.

These bugs are particularly nasty because they are experienced semi-rarely
and under pretty specific circumstances (they only occur when `require_stdlib`
loads another copy of a stdlib, often in a particular order and/or with
a particular state of your pre-compile / loading cache) so they may make
it a long way through a pre-release cycle without an actionable bug
report.