cli: Add `--experimental` flag to enable in-development features
The intention is to use this for "early alpha" kind of features, which are
available to early adopters but still in active feature development (i.e.
they may be very incomplete) and not ready for use in production yet.

The only feature gated behind this right now is `--trim`, whose core
functionality is implemented but which still has a ways to go before we
think the ecosystem / feature-set is ready for general users to adopt.