Fix implicit `convert(String, ...)` in several places
This code was already written very carefully. However, it assumed that
`for x::T in vec` is type-asserting, but it is actually implicitly
converting (what a painful subtlety)

This removes several `convert(String, ...)` from this code, which really
shouldn't be something we invalidate on in the first place but this is
still an improvement in code quality so let's take it.