Split reflection into compiler-dependent and compiler-independent pieces
The `reflection.jl` file provides a large amount of functionality covering
everything from helpers for access to core runtime data structures to setting
up particular inference problems. It is included by both Base and Core.Compiler,
but the functions that use the compiler, don't really make sense in the latter.
In preparation for #56128, and stop including the compiler-dependent pieces in
Core.Compiler.