fix precompile process flags
CacheFlags could get set, but were never propagated to the target
process, so the result would be unusable. Additionally, the debug and
optimization levels were not synchronized with the sysimg, causing a
regression in pkgimage usability after moving out stdlibs. The printing
also failed to meaningfully represent the settings for the same reasons.

Also fixes a concurrency bug in loading.jl tests that was failing.

Fixes #56207
Fixes #56054
Fixes #56206