rename: invalid -> incompatible cache header (#56240)
Falling back to the older serial precompilation process is basically a
bug (except for if a manifest hasn't been resolved) so
https://github.com/JuliaLang/julia/pull/52619 added more info on why
it's been hit so we have a chance of fixing issues that are otherwise
very difficult to recreate.

However "invalid header" which usually just means it was made by a
different julia version appears to sound too alarming to users.
https://discourse.julialang.org/t/cache-misses-when-using-packages-since-upgrading-to-1-11/121445

So soften it there and in error messages, given it seems a better
description.

Suggested by @giordano in
https://discourse.julialang.org/t/cache-misses-when-using-packages-since-upgrading-to-1-11/121445/4?u=ianshmean

(cherry picked from commit aa51abe6400bc29c6093dca5a395fc03806ab511)