Fix kron indexing for types without a unique zero (#56229)
This fixes a bug introduced in
https://github.com/JuliaLang/julia/pull/55941.

We may also take this opportunity to limit the scope of the `@inbounds`
annotations, and also use `axes` to compute the bounds instead of
hard-coding them.

The real "fix" here is on line 767, where `l in 1:nA` should have been
`l in 1:mB`. Using `axes` avoids such errors, and makes the operation
safer as well.