lowering: split `finally` blocks for exceptional control-flow (#55876)
This change duplicates `finally` blocks in lowered IR, so that they can
have a static nesting depth in the `try-catch` hierarchy.

Previously, `finally` control-flow looked like this:

```
error   non-error
    \   /
     \ /
      |
    finally block
      |
     / \
    /   \
 error   non-error
```

This kind of flow is a problem, because in a couple places the compiler
assumes that it can actually "color" the CFG such that there is a static
nesting depth at each BasicBlock (i.e. each BasicBlock can be labeled w/
a unique enclosing `try` / `catch` scope). The above `finally` pattern
violates that assumption.

In an upcoming PR, I want to extend the lifetimes of our Event Handlers
(`jl_handler_t`) until the end of a `catch` block (rather than the
start) which noticeably breaks `llvm-lower-handlers.cpp`. (@keno was
very clear about this assumption in the comments for that pass.)

Behaviorally this was _mostly_ benign, except for some mis-handling of
an erroring entry that turns into a non-erroring exit. That could be fixed
by banning `break` and `return` within `finally` blocks or making the
lowering more complicated, but this PR instead splits the `finally` block
into an erroring and non-erroring path so that we can attach the `catch`
handler appropriately.