Fix fetching parent in symmetric algebra (#56286)
We only need the `parent` of the result if it is a triangular matrix.
For other structurally triangular matrices, such as `Diagonal`s, we may
use these directly in the `Hermitian` constructor.

The operation proceeds as (assuming `H isa Hermitian` with `H.uplo ==
'U'`):
```julia
function +(H::Hermitian, H::Hermitian)
    U = uppertriangular(parent(H))
    Ures = U + U
    data = Ures isa UpperTriangular ? parent(Ures) : Ures # this PR
    Hermitian(data, :U)
end
```
This accounts for the fact that `Ures` may not be an `UpperTriangular`,
as `uppertriangular` might be specialized by the parent. In such cases
we should not extract the parent. Currently, only `Diagonal` specializes
`uppertriangular`, so this issue only exists for a `Hermitian` or a
`Symmetric` that wraps a `Diagonal`.

Fixes https://github.com/JuliaLang/julia/issues/56283

(cherry picked from commit 5e4fb519b5955b2f9ff9a26bd1dc5454561ecef7)