[1.11.x] Revert "yet more atomics & cache-line fixes on work-stealing queue (#53424)"
Results in reliable `julia` segfault immediately upon startup.
I don't know what is going on or why it is crashing or what is wrong,
but this is very reproducible and bisection result is reliable.

Notably, does not seem to be happening in `master`.

See the following issue for more details:
Fixes https://github.com/JuliaLang/julia/issues/54560

This reverts commit 53180e48ad973e5fa479327fb22ff2e1e5dba284.