[backports-release-1.11] Allow ext → ext dependency if triggers are a strict superset (#56368)
This is an effort at a proper workaround (feature?) to resolve
https://github.com/JuliaLang/julia/issues/56204.

For example:
```toml
[extensions]
PlottingExt = "Plots"
StatisticsPlottingExt = ["Plots", "Statistics"]
```

Here `StatisticsPlottingExt` is allowed to depend on `PlottingExt` This
provides a way to declare `ext → ext` dependencies while still avoiding
any extension cycles. The same trick can also be used to make an
extension in one package depend on an extension provided in another.

We'll also need to land #49891, so that we guarantee these load in the
right order.