Add a rendezvous mechanism for reflection functions
This adds a reflection mechanism for the various reflection functions
defined by Base. These functions are mostly interactive affordances and
for ease of Compiler development, it tends to be important that these
reflect the user's expectation of what the compiler is. This is currently
the case using the existing `Revise.track(Core.Compiler)` workflow,
because it runs in the latest world-age, not the inference world.

This commit adds a special binding that if present in `Main` (because
it was imported by a `using Compiler`) will switch the compiler used
for reflection to whatever the binding points to.

In this way, we can replace the old
`using Revise; Revise.track(Core.Compiler)` workflow by
`using Revise, Compiler` without additional headache.