EA: disable finalizer inlining for allocations that are edges of `PhiNode`s
The current EA-based finalizer inlining implementation can create
invalid IR when the target object is later aliased as a `PhiNode`,
which was causing JuliaLang/julia#56422.
In such cases, finalizer inlining for the allocations that are edges of
each `PhiNode` should be avoided, and instead, finalizer inlining should
ideally be applied to the `PhiNode` itself, but implementing that is
somewhat complex. As a temporary fix, this commit disables inlining in
those cases.

- fixes JuliaLang/julia#56422