Undo the decision to publish incomplete types to the binding table
This effectively reverts #36121 and replaces it with #36111, which
was the originally proposed alternative to fix #36104. To recap,
the question is what should happen for
```
module Foo
    struct F
        v::Foo.F
    end
end
```
i.e. where the type reference tries to refer to the newly defined
type via its global path. In #36121 we adjusted things so that we
first assign the type to its global binding and then evaluate
the field type (leaving the type in an incomplete state in the
meantime). The primary reason that this choice was that we would
have to deal with incomplete types assigned to global bindings
anyway if we ever did #32658. However, I think this was the wrong
choice. There is a difference between allowing incomplete types
and semantically forcing incomplete types to be globally observable
every time a new type is defined.

The situation was a little different four years ago, but with more
extensive threading (which can observe the incompletely constructed
type) and the upcoming completion of bindings partition, the
situation is different. For bindings partition in particular, this
would require two invalidations on re-definition, one to the new
incomplete type and then back to the complete type. I don't think
this is worth it, for the (somewhat niche and possibly-should-be-
deprecated-future) case of refering to incompletely defined types
by their global names.

So let's instead try the hack in #36111, which does a frontend
rewrite of the global path. This should be sufficient to at least address
the obvious cases.