Sockets: fix `getipaddr()`
This PR updates the definition of `getipaddr()` so that the
documentation is respected.

Before this change, `getipaddr()` would throw an error unless there is
an `IPv4` address assigned to a network interface. After this change,
`getipaddr()` preferably returns an `IPv4` address, but if none exists,
it returns an `IPv6` address.

Fixes #56520