Fix and test special loading support for Compiler
The new `Compiler` package has a special stub that bypasses
compilig a separate copy if you have `dev`'ed the version
that's already compiled into the sysimg. It wasn't quite
working properly in the final version of that PR if a
compiler so loaded was a dependency of another precompiled
package. Fix that and add a test to make sure it doesn't
regress.