Some misc compiler-related code cleanup
I'm cleaning up some downstream packages following the compiler
split and these were commonly found in compat code. One is
the `Base.copy` method for Phi(C)Node, which we had an implementation
for, just not wired up to `copy`. The other is `block_for_inst`
with an SSAValue, which exists for IncremetalCompact, but not IRCode,
so add the latter for consistency.